+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+### 1.0.2
+
+* Added Document class
+* Patch handles now also RecordInterface objects
+* Improved comparator
+
 ### 1.0.1 (2016-07-10)
 
 * Deprecate WebSignature and WebToken class